Add files via upload
1. Get the code presented in class working. i.e. setup and draw methods, human player
(bat), and ball. The integer variables mouseX and mouseY provide the x and y coordinates of the mouse pointer.
(2 marks)
2. Use the player class to create a computer player. You will need to create a new
method to control the computer's movement. A simple algorithm to use would be to
have the computer player move 1 pixel towards the ball position every frame. The
draw method for your new program will:
 move the ball
 move the computer player
 move the human player
 check for collision with human player
 check for collision with computer player
 draw the ball
 draw the computer player
 draw the human player
(3 marks).
3. Implement a reset() method which is called when the ball falls off the screen
(either the human or the computer has lost the game). The method should put the ball
back onto the screen and restart the game once the mouse button is pressed. If you
define a method called mousePressed(), mousePressed(), Processing will call it when the mouse is
pressed. (3 marks).
4. Count the number of lives (hint: create an attribute and method in player). If the
computer wins three times, a game over message is displayed. If the player wins three
times, a different message is displayed. (2 marks).