Add files via upload
For this lab you may wish to download some files from the CS1013 module on blackboard. It includes:
• Outlines for the program.
• The example array code presented in class
• exploding.gif and invader.gif
1. Create an “Alien” class that has x, y position on the screen, a constructor, a draw method (drawing an
image – “invader.gif”), and a move method. Construct a program which creates an instance of this class,
and moves it across the screen using the move and draw methods. When it reaches the edge of the screen
it should move down smoothly a distance equal to its own height and continue in the opposite direction.
(4 Marks)
For part 1 you should look at Lab 1 for ideas / examples. Think about a “state” in which direction is
recorded, either (Right X ++), (Left X--) or (Down Y++). Maybe use your lab 1 code as a starting point,
rather than overlapping move the square down and reverse its direction when it hits the side.
2. Using the array code presented in class as a template, write a program that creates an array of Alien
objects (10 Aliens in total), initializes the array with new alien objects, and uses the move and draw
methods to move all of the objects across and down the screen. (4 Marks)
For part 2 look at the example circle code presented in the Lecture. You are using an array to store
multiple aliens instead of accessing a single Alien.
3. Create an explode method which changes the appearance of the alien (each alien should be able to
change independently). Demonstrate the method by having the aliens explode at random. You can use a
second PImage loaded with exploding.gif (2 Marks)
Think about “status” which is either exploding or not exploding. This can be used when you need to
draw each alien. Also look at the random() processing method.