Add files via upload
1. Building on the Alien class from last week (or the sample code given in the
lecture), and the Player class from week 2, add a Player object to your program so
that the Player can be moved around at the bottom of the screen while the Aliens
move across and down the screen. (4 marks)
2. Add a Bullet class with a move method which moves the Bullet up the screen, and
a draw method which draws the bullet. An instance of the Bullet class should be
created at the Player’s position when the mouse button is pressed. (3 marks)
3. Implement a collide method in the Bullet class to check whether the bullet has
collided with an Alien. The array of Aliens should be passed as an argument to this
method. Demonstrate your code by being able to fire at aliens and have them explode
(e.g. using the image from last week) when hit. Note - a single bullet can hit multiple
aliens, these are futuristic bullets. (3 marks)