Add files via upload
1. Using the widgetList program and Widget class covered at the lecture, write a
program containing three widgets/buttons labelled "red", "green" and "blue". Alter the
mousePressed method so that a square displayed on the screen changes colour (to
the named colour) as each button is pressed. (3 marks)
2. Alter your program so that the border colour of a widget changes (to white) when
the mouse pointer is over it but reverts to normal (black) when the mouse pointer is
not over it. This must work independently for each of your widgets (i.e. only the
widget the mouse over is highlighted like this). Processing will call the mouseMoved
method (if you define it) whenever the mouse is moved. The stroke() method
allows you to define the border color. (3 marks)
3. Define a Screen class which contains an ArrayList of Widgets. Each screen
should have it’s own background colour. Give Screen a getEvent method which
returns an event (the Widget pressed), and define a draw method in Screen which
draws the screen's widgets. Each screen has it’s own background color.
Create an addWidget method to add a widget to the ArrayList of the screen. Use
your Screen class to write a program where there are two screens, and two buttons
on each screen, one of the buttons on each screen should move you forward and back
through the different screens as illustrated below. You can do this by creating two
instances of Screen. Pressing the other button should result in some text output from
println to report which button has been pressed. Hint: you can create an extra
Screen variable currentScreen which keeps track of the current screen – i.e. one of the
existing instances.
Your mousePressed method in the main program will no longer require a loop, but
should still deal with the different events. (4 marks)